libm src fixes

+++ usr.lib/libm/src/s_floor.c
@@ -62,7 +62,7 @@ floor(double x)
 		    if(jj0==20) i0+=1;
 		    else {
 			j = i1+(1<<(52-jj0));
-			if(j<i1) i0 +=1 ; 	/* got a carry */
+			if(j<(u_int32_t)i1) i0 +=1 ; 	/* got a carry */
 			i1=j;
 		    }
 		}
